fov adjusted maxmipdist. but still not correct to size.

+++ src/ShaderConstants.h
@@ -18,6 +18,6 @@ struct PerFrameUBO
     float padding2;
     glm::vec3 uCamUp;
     float padding3;
-    glm::uvec2 uResolution;
+    glm::vec2 uResolution;
     float uTime;
 };
